establish testing syntax
