+++ idea-plugin/src/test/kotlin/arrow/meta/ide/plugins/higherkinds/HigherKindTest.kt
@@ -1,15 +1,20 @@
 package arrow.meta.ide.plugins.higherkinds
 
+import arrow.meta.ide.testing.IdeTest
+import arrow.meta.ide.testing.env.assertThis
+import arrow.meta.ide.testing.source
+import org.junit.Test
+
 class HigherKindTest {
-  /*@Test
+  @Test
   fun `LM Test`(): Unit =
     assertThis(
       IdeTest(
-        code = HigherKindsTestCode.code.source,
+        code = "HigherKindsTestCode.code".source,
         assert = {
           // availableLM(ArrowIcons.HKT, HigherKindsTestCode.code.source)
           empty
         }
       )
-    )*/
+    )
 }
