+++ idea-plugin/src/test/kotlin/arrow/meta/ide/plugins/higherkinds/HigherKindsTestCode.kt
