+++ idea-plugin/src/test/kotlin/arrow/meta/ide/plugins/typeclasses/TypeclassesTestCode.kt
@@ -1,0 +1,2 @@
+package arrow.meta.ide.plugins.typeclasses
+
